Updated standalone to Jetty 6.0.1, fixed installer webapp, updated quartz to 1.6.
git-svn-id: http://javaforge.com/svn/luntbuild/trunk@224 7c8a6376-b613-0410-8ed7-c433467a09df

+++ luntbuild/installer/scripts/configure-war.sh
@@ -10,7 +10,9 @@ cd "$INSTALL_PATH/web"
 jar -cvf "$INSTALL_PATH/luntbuild.war" .
 cd "$INSTALL_PATH"
 if [ "$appServerWebapps" != "" ]; then
+if [ "$appServerWebapps" != "." ]; then
 rm -rf "$appServerWebapps/luntbuild"
 cp -f "$INSTALL_PATH/luntbuild.war" "$appServerWebapps/luntbuild.war"
+fi
 fi
 chmod a+x bin/*
